Add projects for exploring Merkle trees/proofs
The main project file here is `src/stepByStepMerkle.js`. In this file, I
explored building a Merkle root based on a given set of transaction IDs,
building a Merkle proof for a given transaction, and verifying a
transaction using a Merkle proof.

The other files under `src` were used mainly for exploration purposes.